fix: Abstract out datasource functionality from the Scan
